- .NET implementation of entropy uses float now (just as NativeCrypt) - fixed storage key generator in distributed keysearcher
git-svn-id: https://svn.cryptool.org/CrypTool2/trunk@2848 9b0d660e-8ab2-dd11-ad6d-000c29d71a56

+++ CrypPlugins/CostFunction/CostFunction.cs
@@ -556,7 +556,7 @@ namespace Cryptool.Plugins.CostFunction
                         n[text[counter]]++;
                     }
 
-                    double entropy = 0;
+                    float entropy = 0;
                     //calculate probabilities and sum entropy
                     for (int i = 0; i < 256; i++)
                         entropy += xlogx[n[i]];
